fix: renamed module to avoid import conflicts
